Separate path and manifest helpers into sub-namespaces
